+++ compiler/src/org.graalvm.compiler.truffle.common/src/org/graalvm/compiler/truffle/common/TruffleSourceLanguagePosition.java
@@ -43,4 +43,6 @@ public interface TruffleSourceLanguagePosition {
     int getNodeId();
 
     String getNodeClassName();
+
+    String getQualifiedRootName();
 }