+++ compiler/src/org.graalvm.compiler.truffle.compiler/src/org/graalvm/compiler/truffle/compiler/PartialEvaluator.java
@@ -574,5 +574,10 @@ public abstract class PartialEvaluator {
             return delegate.getNodeClassName();
         }
 
+        @Override
+        public String getQualifiedRootName() {
+            return delegate.getQualifiedRootName();
+        }
+
     }
 }